feat: add chrono/timespec/timeval interop for time types
Add si_chrono.hpp with constexpr noexcept conversions between
strong time types and std::chrono::duration, struct timespec,
and struct timeval. Also adds Nanoseconds<T> alias, _ns UDL,
and fixes pre-existing GCC 15 warnings in test files.

+++ test/strong_aligned_array_test.cpp
@@ -106,7 +106,7 @@ constexpr Vec make_seq(float start, float step = 1.0f)
     Vec out{};
     for (std::size_t i = 0; i < 8; ++i)
     {
-        out[i] = start + i * step;
+        out[i] = start + static_cast<float>(i) * step;
     }
     return out;
 }
